requirements.txt
streamlit
youtube-transcript-api
pandas
openai
matplotlib
wordcloud
load_dotenv